# backend-go/docs/swagger.yaml
basePath: /api
definitions:
...
  presentation.BookingResponse:
    properties:
      createdAt:
        type: string
      endTime:
        type: string
      id:
        type: integer
      notes:
        type: string
      paymentStatus:
        type: string
      pistaId:
        type: integer
      pistaName:
        type: string
      pistaType:
        type: string
      priceSnapshotCents:
        type: integer
      priceSnapshotEuros:
        description: Convertido a euros para UI
        type: number
      startTime:
        type: string
      status:
        type: string
      updatedAt:
        type: string
      userId:
        type: integer
      userName:
        type: string
    type: object
  presentation.ClassResponse:
    properties:
      capacity:
        type: integer
      createdAt:
        type: string
      description:
        type: string
      endTime:
        type: string
      enrolledCount:
        type: integer
      enrollments:
        items:
          $ref: '#/definitions/presentation.EnrollmentResponse'
        type: array
      id:
        type: integer
      instructorId:
        type: integer
      instructorName:
        type: string
      pistaId:
        type: integer
      pistaName:
        type: string
      priceCents:
        type: integer
      priceEuros:
        type: number
      slug:
        type: string
      startTime:
        type: string
      status:
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
...
  presentation.CreateBookingRequest:
    properties:
      endTime:
        type: string
      notes:
        type: string
      pistaId:
        minimum: 1
        type: integer
      startTime:
        type: string
      userId:
        type: integer
    required:
    - endTime
    - pistaId
    - startTime
    type: object
...
  presentation.CreateClassRequest:
    properties:
      description:
        type: string
      endTime:
        type: string
      instructorId:
        type: integer
      maxCapacity:
        maximum: 20
        minimum: 1
        type: integer
      pistaId:
        type: integer
      priceCents:
        minimum: 0
        type: integer
      startTime:
        type: string
      title:
        minLength: 3
        type: string
    required:
    - endTime
    - instructorId
    - maxCapacity
    - pistaId
    - priceCents
    - startTime
    - title
    type: object
...
  presentation.EnrollmentResponse:
    properties:
      classId:
        type: integer
      className:
        type: string
      enrolledAt:
        description: Alias para compatibilidad
        type: string
      id:
        type: integer
      registeredAt:
        type: string
      status:
        type: string
      userEmail:
        type: string
      userId:
        type: integer
      userName:
        type: string
    type: object
...
  presentation.UpdateBookingRequest:
    properties:
      endTime:
        type: string
      notes:
        type: string
      paymentStatus:
        type: string
      pistaId:
        minimum: 1
        type: integer
      startTime:
        type: string
      status:
        type: string
      userId:
        type: integer
    required:
    - endTime
    - pistaId
    - startTime
    type: object
  presentation.UpdateClassRequest:
    properties:
      description:
        type: string
      endTime:
        type: string
      instructorId:
        type: integer
      maxCapacity:
        maximum: 20
        minimum: 1
        type: integer
      pistaId:
        type: integer
      priceCents:
        minimum: 0
        type: integer
      startTime:
        type: string
      status:
        type: string
      title:
        minLength: 3
        type: string
    required:
    - endTime
    - instructorId
    - maxCapacity
    - pistaId
    - priceCents
    - startTime
    - title
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API REST para gestión de instalaciones deportivas
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: PoliManage API
  version: "1.0"
paths:
  /api/bookings:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/presentation.BookingResponse'
            type: array
      summary: Listar todas las reservas
      tags:
      - bookings
    post:
      consumes:
      - application/json
      parameters:
      - description: Datos de la reserva
        in: body
        name: booking
        required: true
        schema:
          $ref: '#/definitions/presentation.CreateBookingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/presentation.BookingResponse'
      summary: Crear una nueva reserva
      tags:
      - bookings
  /api/bookings/{id}:
    delete:
      parameters:
      - description: ID de la reserva
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: Eliminar una reserva
      tags:
      - bookings
    get:
      parameters:
      - description: ID de la reserva
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presentation.BookingResponse'
      summary: Obtener reserva por ID
      tags:
      - bookings
    put:
      consumes:
      - application/json
      parameters:
      - description: ID de la reserva
        in: path
        name: id
        required: true
        type: integer
      - description: Datos a actualizar
        in: body
        name: booking
        required: true
        schema:
          $ref: '#/definitions/presentation.UpdateBookingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presentation.BookingResponse'
      summary: Actualizar una reserva
      tags:
      - bookings
  /api/bookings/{id}/cancel:
    post:
      parameters:
      - description: ID de la reserva
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presentation.BookingResponse'
      summary: Cancelar una reserva
      tags:
      - bookings
  /api/bookings/pista/{pistaId}/date/{date}:
    get:
      parameters:
      - description: ID de la pista
        in: path
        name: pistaId
        required: true
        type: integer
      - description: Fecha (YYYY-MM-DD)
        in: path
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/presentation.BookingResponse'
            type: array
      summary: Obtener reservas por pista y fecha
      tags:
      - bookings
  /api/classes:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/presentation.ClassResponse'
            type: array
      summary: Listar todas las clases
      tags:
      - classes
    post:
      consumes:
      - application/json
      parameters:
      - description: Datos de la clase
        in: body
        name: class
        required: true
        schema:
          $ref: '#/definitions/presentation.CreateClassRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/presentation.ClassResponse'
      summary: Crear una nueva clase
      tags:
      - classes
  /api/classes/{slug}:
    delete:
      parameters:
      - description: Slug de la clase
        in: path
        name: slug
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Eliminar una clase
      tags:
      - classes
    get:
      parameters:
      - description: Slug de la clase
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presentation.ClassResponse'
      summary: Obtener clase por slug
      tags:
      - classes
    put:
      consumes:
      - application/json
      parameters:
      - description: Slug de la clase
        in: path
        name: slug
        required: true
        type: string
      - description: Datos a actualizar
        in: body
        name: class
        required: true
        schema:
          $ref: '#/definitions/presentation.UpdateClassRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presentation.ClassResponse'
      summary: Actualizar una clase
      tags:
      - classes
  /api/classes/{slug}/cancel:
    post:
      parameters:
      - description: Slug de la clase
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presentation.ClassResponse'
      summary: Cancelar una clase
      tags:
      - classes
...
  /api/classes/{slug}/enrollments:
    get:
      parameters:
      - description: ID de la clase
        in: query
        name: classID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/presentation.EnrollmentResponse'
            type: array
      summary: Obtener inscripciones de una clase
      tags:
      - enrollments
  /api/classes/instructor/{instructorId}:
    get:
      parameters:
      - description: ID del instructor
        in: path
        name: instructorId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/presentation.ClassResponse'
            type: array
      summary: Obtener clases por instructor
      tags:
      - classes
...
schemes:
- http
swagger: "2.0"